requirements.txt
pytest
xmltodict
pytest-cov
dictdiffer
selenium
webdriver_manager
allure-pytest
requests
